requirements.txt
# Core
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# HTTP & Scraping
httpx>=0.26.0
playwright>=1.41.0
beautifulsoup4>=4.12.0
parsel>=1.8.0

# Notifications
aiosmtplib>=3.0.0

# UI
streamlit>=1.31.0
pandas>=2.2.0

# Scheduling & Utils
apscheduler>=3.10.0
rapidfuzz>=3.6.0
structlog>=24.1.0
tenacity>=8.2.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
